# docs/swagger.yaml
basePath: /api
definitions:
  models.APIError:
    description: Estrutura de erro padrão retornada pela API
    properties:
      code:
        example: 400
        type: integer
      details:
        example: O campo error_details é obrigatório
        type: string
      message:
        example: Parâmetros inválidos
        type: string
    required:
    - code
    - message
    type: object
  models.ErrorRequest:
    description: Requisição contendo os detalhes do erro a ser analisado
    properties:
      context:
        example: Deployment em cluster Kubernetes 1.26 com imagem Docker personalizada
        type: string
      error_details:
        example: 'CrashLoopBackOff: container failed to start'
        type: string
    required:
    - error_details
    type: object
  models.ErrorResponse:
    description: Resposta contendo análise e solução para o erro reportado
    properties:
      error:
        $ref: '#/definitions/models.ErrorSolution'
      message:
        example: Análise concluída com sucesso
        type: string
    required:
    - error
    type: object
  models.ErrorSolution:
    description: Estrutura contendo a causa identificada e soluções propostas para
      o erro
    properties:
      causa:
        example: Imagem Docker inválida
        type: string
      solucao:
        example: |-
          kubectl describe pod meu-pod
          kubectl logs meu-pod --previous
        type: string
    required:
    - causa
    - solucao
    type: object
host: localhost:8080
info:
  contact: {}
  description: Error resolution API using local LLM
  title: Hefestus API
  version: "1.0"
paths:
  /errors/{domain}:
    post:
      consumes:
      - application/json
      description: Recebe detalhes de um erro e seu contexto, retornando possíveis
        soluções baseadas em LLM
      parameters:
      - description: Domínio técnico (kubernetes, github, argocd)
        enum:
        - kubernetes
        - github
        - argocd
        in: path
        name: domain
        required: true
        type: string
      - description: Detalhes do erro e contexto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ErrorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Solução para o erro
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "400":
          description: Erro de validação ou requisição inválida
          schema:
            $ref: '#/definitions/models.APIError'
        "404":
          description: Domínio não encontrado
          schema:
            $ref: '#/definitions/models.APIError'
        "500":
          description: Erro interno do servidor
          schema:
            $ref: '#/definitions/models.APIError'
      summary: Analisar e resolver erros por domínio
      tags:
      - errors
schemes:
- http
swagger: "2.0"
tags:
- description: Error resolution endpoints
  name: errors